+++ packages/atomic/src/components/facets/atomic-numeric-facet/atomic-numeric-facet.tsx
@@ -137,6 +137,16 @@ export class AtomicNumericFacet
    * Specifies if the facet is collapsed.
    */
   @Prop({reflect: true, mutable: true}) public isCollapsed = false;
+  /**
+   * Whether to exclude the parents of folded results when estimating the result count for each facet value.
+   */
+  @Prop() public filterFacetCount = true;
+  /**
+   * The maximum number of results to scan in the index to ensure that the facet lists all potential facet values.
+   * Note: A high injectionDepth may negatively impact the facet request performance.
+   * Minimum: `0`
+   */
+  @Prop() public injectionDepth = 1000;
 
   private validateProps() {
     new Schema({
@@ -165,6 +175,8 @@ export class AtomicNumericFacet
       rangeAlgorithm: this.rangeAlgorithm,
       currentValues: this.manualRanges,
       generateAutomaticRanges: !this.manualRanges.length,
+      filterFacetCount: this.filterFacetCount,
+      injectionDepth: this.injectionDepth,
     };
     this.facet = buildNumericFacet(this.bindings.engine, {options});
     this.facetId = this.facet.state.facetId;