+++ packages/atomic/src/components/result-template-components/atomic-result-icon/icons.d.ts
@@ -1,0 +1,4 @@
+declare module '@salesforce-ux/design-system/design-tokens/dist/bg-standard.common' {
+  const icons: Record<string, string>;
+  export default icons;
+}