+++ packages/quantic/force-app/main/default/lwc/quanticNumberButton/quanticNumberButton.js
@@ -26,7 +26,7 @@ export default class QuanticNumberButton extends LightningElement {
   }
 
   get variant() {
-    return this.number === this.current ? 'brand-outline' : 'brand';
+    return this.number === this.current ? 'brand' : 'brand-outline';
   }
 
   goto() {